+++ package.json
@@ -1,8 +1,8 @@
 {
-  "name": "new-project",
-  "version": "1.0.0",
+  "name": "cyber_market",
+  "version": "1.2.0",
   "private": true,
-  "description": "new-project descr",
+  "description": "Project of web store",
   "scripts": {
     "build": "webpack",
     "start": "webpack-dev-server --open"
