+++ public/html/pageProduct.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>pageProduct</title>
+    <title>Page Product</title>
     <link rel="stylesheet" href="./style/pageProduct.css">
     <link rel="stylesheet" href="./fontawesome/css/all.css">
 </head>
