+++ public/html/personalCabinet.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>Cabinet</title>
+    <title>Personal Cabinet</title>
     <link rel="stylesheet" href="./style/styleCbinet.css">
     <link rel="stylesheet" href="./fontawesome/css/all.css">
 </head>
